tests: Rename test source filenames and executables
